fixed the galaxy shader rendering
it was a simple c++ and hlsl struct content type mismatch issue

+++ src/core/Buffer.h
@@ -51,7 +51,7 @@ struct DepthPPBuffer
 
 struct ShaderEffects
 {
-    u32 _resolution[2];
+    float _resolution[2];
     float _time;
     float _cameraYaw;
